+++ sentry/src/test/java/io/sentry/HubTest.kt
@@ -1407,7 +1407,7 @@ class HubTest {
     fun `when startTransaction and profiling is enabled, transaction is profiled only if sampled`() {
         val mockTransactionProfiler = mock<ITransactionProfiler>()
         val mockClient = mock<ISentryClient>()
-        whenever(mockTransactionProfiler.onTransactionFinish(any())).thenAnswer { mockClient.captureEnvelope(mock()) }
+        whenever(mockTransactionProfiler.onTransactionFinish(any(), anyOrNull())).thenAnswer { mockClient.captureEnvelope(mock()) }
         val hub = generateHub {
             it.setTransactionProfiler(mockTransactionProfiler)
         }
@@ -1429,7 +1429,7 @@ class HubTest {
     fun `when startTransaction and is sampled but profiling is disabled, transaction is not profiled`() {
         val mockTransactionProfiler = mock<ITransactionProfiler>()
         val mockClient = mock<ISentryClient>()
-        whenever(mockTransactionProfiler.onTransactionFinish(any())).thenAnswer { mockClient.captureEnvelope(mock()) }
+        whenever(mockTransactionProfiler.onTransactionFinish(any(), anyOrNull())).thenAnswer { mockClient.captureEnvelope(mock()) }
         val hub = generateHub {
             it.profilesSampleRate = 0.0
             it.setTransactionProfiler(mockTransactionProfiler)